pyproject.toml
[tool.poetry]
name = "diplomacy-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Patrick <pakue@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = {extras = ["job-queue"], version = "^20.7"}
numpy = "^1.26.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
